fix compiler warnings
git-svn-id: https://svn.mi.fu-berlin.de/des-testbed/Software/DES-SERT/des-olsr/trunk@8375 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ src/pipeline/olsr_pipeline.c
@@ -36,7 +36,7 @@ pthread_rwlock_t rlseqlock = PTHREAD_RWLOCK_INITIALIZER;
 
 // ---------------------------- pipeline callbacks ---------------------------------------------
 
-dessert_cb_result olsr_drop_errors(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result olsr_drop_errors(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     if(proc->lflags & DESSERT_RX_FLAG_L2_SRC
        || proc->lflags & DESSERT_RX_FLAG_L25_SRC) {
         struct ether_header* l25h = dessert_msg_getl25ether(msg);
@@ -51,7 +51,7 @@ dessert_cb_result olsr_drop_errors(dessert_msg_t* msg, size_t len, dessert_msg_p
     return DESSERT_MSG_KEEP;
 }
 
-dessert_cb_result olsr_handle_hello(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result olsr_handle_hello(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     dessert_ext_t* ext;
 
     if(dessert_msg_getext(msg, &ext, HELLO_EXT_TYPE, 0) != 0) {
@@ -209,7 +209,7 @@ dessert_cb_result olsr_handle_hello(dessert_msg_t* msg, size_t len, dessert_msg_
     return DESSERT_MSG_KEEP;
 }
 
-dessert_cb_result olsr_handle_tc(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result olsr_handle_tc(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     dessert_ext_t* ext;
 
     if(dessert_msg_getext(msg, &ext, TC_EXT_TYPE, 0) != 0) {
@@ -278,7 +278,7 @@ dessert_cb_result olsr_handle_tc(dessert_msg_t* msg, size_t len, dessert_msg_pro
     return DESSERT_MSG_KEEP;
 }
 
-dessert_cb_result olsr_handle_ett(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result olsr_handle_ett(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     dessert_ext_t* ext;
 
     if(dessert_msg_getext(msg, &ext, ETT_EXT_TYPE, 0) != 0) {
@@ -333,7 +333,7 @@ dessert_cb_result olsr_handle_ett(dessert_msg_t* msg, size_t len, dessert_msg_pr
     return DESSERT_MSG_KEEP;
 }
 
-dessert_cb_result olsr_fwd2dest(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result olsr_fwd2dest(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     struct ether_header* l25h = dessert_msg_getl25ether(msg);
     dessert_ext_t* rl_ext;
     uint32_t rl_seq_num = 0;
@@ -430,7 +430,7 @@ dessert_cb_result olsr_fwd2dest(dessert_msg_t* msg, size_t len, dessert_msg_proc
 
 // --------------------------- sysif handling ----------------------------------------------------------
 
-dessert_cb_result olsr_sys2rp(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_sysif_t* tunif, dessert_frameid_t id) {
+dessert_cb_result olsr_sys2rp(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_sysif_t* tunif, dessert_frameid_t id) {
     struct ether_header* l25h = dessert_msg_getl25ether(msg);
 
     // L25 destination is broadcast; TODO what about multicast?
@@ -481,7 +481,7 @@ dessert_cb_result olsr_sys2rp(dessert_msg_t* msg, size_t len, dessert_msg_proc_t
 /**
 * Forward packets addressed to me via sysif. The packet is always dropped after handling.
 */
-dessert_cb_result rp2sys(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
+dessert_cb_result rp2sys(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id) {
     if((proc->lflags & DESSERT_RX_FLAG_L25_DST && !(proc->lflags & DESSERT_RX_FLAG_L25_OVERHEARD))
        || proc->lflags & DESSERT_RX_FLAG_L25_BROADCAST
        || proc->lflags & DESSERT_RX_FLAG_L25_MULTICAST) {
